I finished Task3
In this one, I discovered a mistake that may have been in the Task 1
and 2, I will fix those too in the next commit.  This one has the FSM
design, and the junit tests working correctly.  Also, I had already
named the project “Project5” for Task1 and Task 2. So the project name
for task 3 is “Project5Task3”.